+++ src/main/java/com/vowing/parang/domain/category/dto/CategoryDTO.java
@@ -1,0 +1,13 @@
+package com.vowing.parang.domain.category.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class CategoryDTO {
+    private Long id;
+    private String category;
+}
