+++ src/main/java/com/vowing/parang/domain/category/repository/CategoryRepository.java
@@ -1,0 +1,11 @@
+package com.vowing.parang.domain.category.repository;
+
+import com.vowing.parang.domain.category.entity.CategoryEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CategoryRepository extends JpaRepository<CategoryEntity, Long> {
+
+    CategoryEntity findByCategory(String category);
+}
