+++ src/main/java/com/vowing/parang/domain/log/error/LogNotFound.java
@@ -1,0 +1,7 @@
+package com.vowing.parang.domain.log.error;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(code = HttpStatus.NOT_FOUND)
+public class LogNotFound extends RuntimeException {}
