+++ src/main/java/com/vowing/parang/domain/member/dto/MemberDTO.java
@@ -1,0 +1,16 @@
+package com.vowing.parang.domain.member.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class MemberDTO {
+    private Long id;
+    private String userId;
+    private String password;
+    private String memo;
+}
