+++ src/main/java/com/vowing/parang/domain/member/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package com.vowing.parang.domain.member.repository;
+
+import com.vowing.parang.domain.member.entity.MemberEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
+    Optional<MemberEntity> findByUserId(String userId);
+}
