--- src/main/java/com/vowing/parang/dto/CategoryDTO.java
@@ -1,13 +1,0 @@
-package com.vowing.parang.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class CategoryDTO {
-    private Long id;
-    private String category;
-}
