--- src/main/java/com/vowing/parang/dto/MemberDTO.java
@@ -1,16 +1,0 @@
-package com.vowing.parang.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class MemberDTO {
-    private Long id;
-    private String userId;
-    private String password;
-    private String memo;
-}
