--- src/main/java/com/vowing/parang/entity/CategoryEntity.java
@@ -1,38 +1,0 @@
-package com.vowing.parang.entity;
-
-import com.vowing.parang.dto.CategoryDTO;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Table(name = "Category")
-public class CategoryEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "category", nullable = false)
-    private String category;
-
-    public CategoryDTO toValueObject() {
-        return new CategoryDTO(
-                this.getId(),
-                this.getCategory()
-        );
-    }
-}
