--- src/main/java/com/vowing/parang/error/LogNotFound.java
@@ -1,7 +1,0 @@
-package com.vowing.parang.error;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(code = HttpStatus.NOT_FOUND)
-public class LogNotFound extends RuntimeException {}
