--- src/main/java/com/vowing/parang/error/UserNotFound.java
@@ -1,10 +1,0 @@
-package com.vowing.parang.error;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-/**
- * 사용자를 찾을 수 없음
- */
-@ResponseStatus(code = HttpStatus.NOT_FOUND)
-public class UserNotFound extends RuntimeException {}
