--- src/main/java/com/vowing/parang/repository/CategoryRepository.java
@@ -1,11 +1,0 @@
-package com.vowing.parang.repository;
-
-import com.vowing.parang.entity.CategoryEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CategoryRepository extends JpaRepository<CategoryEntity, Long> {
-
-    CategoryEntity findByCategory(String category);
-}
