--- src/main/java/com/vowing/parang/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package com.vowing.parang.repository;
-
-import com.vowing.parang.entity.MemberEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
-    Optional<MemberEntity> findByUserId(String userId);
-}
